zcash_protocol: Expose transaction version constants.
This eliminates duplication of these constants in the `pczt` module.
It introduces a non-optional dependency of `pczt` on `zcash_protocol`,
but as the `pczt` crate is almost not meaningfully useful without the
`zcash_protocol` dependency anyway, and `zcash_protocol` is very
lightweight, this is an acceptable tradeoff.

(cherry picked from commit aff4e42afaa00c2e08d224863ea064564033d208)